+++ tests/test_utils.py
@@ -118,30 +118,3 @@ class TestRangeHeaders(unittest.TestCase):
         for header in headers:
             res = parse_range_header(header['header'])
             self.assertEqual(res, header['result'])
-
-
-class TestRenderers(unittest.TestCase):
-    '''
-    Renderers have been moved from pyramid_skosprovider.utils to pyramid_skosprovider.renderers.
-    Maintain tests until 0.8.0 when they will be removed.
-    '''
-
-    def test_import_concept_adapter(self):
-        from pyramid_skosprovider.utils import concept_adapter as cau
-        from pyramid_skosprovider.renderers import concept_adapter as car
-        assert cau == car
-
-    def test_import_collection_adapter(self):
-        from pyramid_skosprovider.utils import collection_adapter as cau
-        from pyramid_skosprovider.renderers import collection_adapter as car
-        assert cau == car
-
-    def test_import_source_adapter(self):
-        from pyramid_skosprovider.utils import source_adapter as sau
-        from pyramid_skosprovider.renderers import source_adapter as sar
-        assert sau == sar
-
-    def test_import_json_renderer(self):
-        from pyramid_skosprovider.utils import json_renderer as jru
-        from pyramid_skosprovider.renderers import json_renderer as jrr
-        assert jru == jrr